added comments to allow for easier understanding of friendship tab data loading logic. Renamed some data loading functions to be more descriptive.
